Relocate remoteloader et al into header
