+++ modules/remotebackend/remotebackend.cc
@@ -873,34 +873,4 @@ DNSBackend *RemoteBackend::maker()
    };
 }
 
-class RemoteBackendFactory : public BackendFactory
-{
-  public:
-      RemoteBackendFactory() : BackendFactory("remote") {}
-
-      void declareArguments(const std::string &suffix="")
-      {
-          declare(suffix,"dnssec","Enable dnssec support","no");
-          declare(suffix,"connection-string","Connection string","");
-      }
-
-      DNSBackend *make(const std::string &suffix="")
-      {
-         return new RemoteBackend(suffix);
-      }
-};
-
-class RemoteLoader
-{
-   public:
-      RemoteLoader()
-      {
-#ifdef REMOTEBACKEND_HTTP
-         curl_global_init(CURL_GLOBAL_ALL);
-#endif
-         BackendMakers().report(new RemoteBackendFactory);
-         L<<Logger::Notice<<kBackendId<<" This is the remotebackend version "VERSION" ("__DATE__", "__TIME__") reporting"<<endl;
-      }
-};
-
-RemoteLoader remoteloader;
+RemoteLoader remoteloader __attribute__((visibility("default")));
