+++ modules/remotebackend/test-remotebackend-pipe.cc
@@ -22,18 +22,14 @@
 #include "pdns/statbag.hh"
 #include "pdns/packetcache.hh"
 
+#include "remotebackend.hh"
+
 StatBag S;
 PacketCache PC;
 ArgvMap &arg()
 {
   static ArgvMap arg;
   return arg;
-};
-
-class RemoteLoader
-{
-   public:
-      RemoteLoader();
 };
 
 DNSBackend *be;
@@ -44,7 +40,7 @@ struct RemotebackendSetup {
 	try {
 		// setup minimum arguments
 		::arg().set("module-dir")="";
-                new RemoteLoader();
+		new RemoteLoader();
 		BackendMakers().launch("remote");
                 // then get us a instance of it 
                 ::arg().set("remote-connection-string")="pipe:command=unittest_pipe.rb";
